backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.12.0
python-jose==3.3.0
pydantic==2.6.1
python-dotenv==1.0.1
opencv-python==4.9.0.80
numpy==1.26.4
moviepy==1.0.3
whisper==1.1.10
torch==2.2.0
transformers==4.37.2
sentence-transformers==2.5.1
faiss-cpu==1.7.4
Pillow==10.2.0
PyPDF2 